+++ src/HookVariableInitializer.ts
@@ -2,7 +2,7 @@ import { join, normalize, resolve } from 'path';
 import { writeFileSync } from 'fs';
 import { env } from 'process';
 import { exec } from 'shelljs';
-import HookManager from './HookManager';
+import TempDirectory from './TempDirectory';
 
 export default class HookVariableInitializer {
   public static initEnvVariables(repoPath: string, hookType: string, hookArguments: string[]) {
@@ -20,7 +20,7 @@ export default class HookVariableInitializer {
   }
 
   protected static initPreCommitVars() {
-    const stagedFilesPath = join(HookManager.getTempDirectory(), 'staged-files.json');
+    const stagedFilesPath = join(TempDirectory.getPath(), 'staged-files.json');
     writeFileSync(stagedFilesPath, HookVariableInitializer.getStagedFilesJson());
 
     env.GMH_STAGED_FILES = stagedFilesPath;
@@ -31,11 +31,15 @@ export default class HookVariableInitializer {
       'git diff --cached --name-only --diff-filter=ACM',
       { silent: true },
     );
-    const stagedFiles = rawStagedFiles
-      .slice(0, -1)
-      .split('\n')
-      .map(file => resolve(normalize(file)));
+    if (rawStagedFiles.length > 0) {
+      const stagedFiles = rawStagedFiles
+        .slice(0, -1)
+        .split('\n')
+        .map(file => resolve(normalize(file)));
 
-    return JSON.stringify(stagedFiles);
+      return JSON.stringify(stagedFiles);
+    }
+
+    return JSON.stringify([]);
   }
 }